arm64: Removed PAGE_S2_XN from PAGE_S2
By default in mainline kernel, KVM flags the XN bit on a new page
allocated to the VM. KVM later clears the XN bit when a guest page
is used as executable by checking if the fault was an IABT.

This sadly breaks the grant/revoke memory model in HypSec which
a S2 page table entry could be purged by KVM whenever such the
aforementioned update happens (due to tlb_vmid_ipa which in turns
calls to EL2 as to synchronize updates from soft to shadow S2PT).

We removed PAGE_S2_XN from PAGE_S2 to avoid the page table updates
so to keep the access pattern for memory access control. We may
need to think of something better in the future.