stage2: Replaced enable_stage2_translation with hvc call wrapper
We don't need to write our own assembly code wrapper for
enable_stage2_translation; instead, we can reuse the existing
__hvc_call_hyp wrapper to make hypercalls from host.

We let the host to pass a hypercall number to corevisor, which
it later figures on the corresponding service routine using the
number. Enabling stage2 translation is done in C code now.