stage2: arm64: Avoid clearing S2PT in TLB flushe
Since we now modified PAGE_S2 to not trap on execution, we do not
have to clear S2PT during TLB flushes anymore. This used to keep
soft S2PT in sync with shadow S2PT. It is different story if we'd
like to support swapping or more advanced features that the host
may modify mappings in S2PT for legit reasons.