stage2: arm64: Do not set vmid to 0 in initial grant
Here we don't set vmid of a s2_page to 0 when it is first granted
by a VM, but simply inccrement the counter of the respective
s2_page. Later in handle_host_fault, we reinvent a way to check if
a s2_page is valid by checking if (vmid > 0 && count > 0).